requirements.txt
fastapi
SQLAlchemy
uvicorn
psycopg2-binary
python-dotenv
pytest-cov
pytest
httpx
tqdm